Remove taskList state form Boards to TaskList component
